+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added/Changed/Deprecated/Removed/Fixed/Security: YOUR CHANGE HERE
 
 * Added: error backtrace to SystemError
+* Fixed: skip "base" field name in validation error messages
 
 ## [2.2.0](2022-01-25)